add authorize ro controllers

+++ appet-app/app/Http/Controllers/AppointmentsController.php
@@ -16,6 +16,9 @@ class AppointmentsController extends Controller
      */
     public function index()
     {
+
+        $this->authorize('view-appointments');
+
         return view('appointments.appointments');
     }
 
@@ -26,6 +29,8 @@ class AppointmentsController extends Controller
      */
     public function create()
     {
+        $this->authorize('create-pet');
+
         $user = auth()->user();
         $pets = $user->pets;
 
@@ -40,6 +45,8 @@ class AppointmentsController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create-pet');
+
         $appointments = new Appointment;
 
         $appointments->user_id = auth()->user()->id;
@@ -64,6 +71,8 @@ class AppointmentsController extends Controller
      */
     public function show()
     {
+        $this->authorize('view-appointments');
+
         $user = auth()->user();
         $appointment = $user->appointments;
         //dd($user->appointments);
@@ -78,6 +87,8 @@ class AppointmentsController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('edit-appointment');
+
         $user = auth()->user();
         $pets = $user->pets;
         $appointment = Appointment::findOrFail($id);
@@ -99,6 +110,8 @@ class AppointmentsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('edit-appointment');
+
         $data = $request->all();
         // dd($data);
         // dd(Appointment::findOrFail($request->id));
@@ -114,6 +127,8 @@ class AppointmentsController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete-appointment');
+
         Appointment::findOrFail($id)->delete();
         return redirect('/appointments/show');
     }
